solution commit

+++ src/level14/level14_1_2747.java
@@ -11,10 +11,10 @@ public class level14_1_2747 {
 		
 		Scanner sc = new Scanner(System.in);
 		
-		long input = sc.nextInt(); //입력값
-		long pre = 0;
-		long temp = 0;
-		long result = 1; //결과
+		int input = sc.nextInt(); //입력값
+		int pre = 0;
+		int temp = 0;
+		int result = 1; //결과
 		
 		for(int i=0; i<input-1; i++) {
 			temp = result;
